% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{brute_force_knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{, the dataset}

\item{W}{, the capacity of the knapsack}

\item{parallel, }{make comination using parallel or not
return list , returns a list that contain the maximum value and the list of elements chosen}
}
\description{
brute_force_knapsack function to solve the knapsack problem
}
