% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Knapsack.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{knapsack_dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{, the dataset}

\item{W}{, the capacity of the knapsack
return list , returns a list that contain the maximum value and the list of elements chosen}
}
\description{
knapsack_dynamic function to solve the knapsack problem
}
