% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{This is data to be included in my package}
\format{An object of class \code{data.frame} with 2000 rows and 2 columns.}
\usage{
knapsack_objects
}
\description{
This is data to be included in my package
}
\references{
\url{https://en.wikibooks.org/wiki/R_Programming/Working_with_data_frames}
}
\author{
Noufal \email{muhno651@student.liu.se}
}
\keyword{data}
