\name{Lab06R-package}
\alias{Lab06R-package}
\alias{Lab06R}
\docType{package}
\title{
\packageTitle{Lab06R}
}
\description{
\packageDescription{Lab06R}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab06R}
\packageIndices{Lab06R}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab06R}

Maintainer: \packageMaintainer{Lab06R}
}

